requirements.txt
Flask>=3.0.0
feedparser>=6.0.11
openai>=1.40.0
python-dotenv>=1.0.1
